+++ Runtime/EntitySystem/IBossBar.cs
@@ -1,0 +1,7 @@
+namespace CrystalFlux.Core
+{
+    public interface IBossBar
+    {
+        void Setup(string bossName, IStatProvider stats);
+    }
+}